backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.35
python-dotenv>=1.0.0
pydantic>=2.9.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
openai>=1.75.0
httpx>=0.28.0
psycopg2-binary>=2.9.9
stripe>=8.0.0
